requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
groq
python-dotenv
pydantic
monai
numpy
nibabel
torch
torchvision
sqlalchemy
sqlmodel